% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000-wrappers.R
\docType{data}
\name{OpendalAio}
\alias{OpendalAio}
\title{Asynchronous operation handle.}
\format{
An object of class \code{Ropendal::OpendalAio__bundle} (inherits from \code{savvy_Ropendal__sealed}) of length 0.
}
\usage{
OpendalAio
}
\description{
Asynchronous operation handle.
}
\keyword{datasets}
