% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000-wrappers.R
\docType{data}
\name{OpendalCredentialProvider}
\alias{OpendalCredentialProvider}
\title{Explicit credential provider.}
\format{
An object of class \code{Ropendal::OpendalCredentialProvider__bundle} (inherits from \code{savvy_Ropendal__sealed}) of length 5.
}
\usage{
OpendalCredentialProvider
}
\description{
Explicit credential provider.
}
\keyword{datasets}
