% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000-wrappers.R
\docType{data}
\name{OpendalFs}
\alias{OpendalFs}
\title{Filesystem handle backed by Apache OpenDAL.}
\format{
An object of class \code{Ropendal::OpendalFs__bundle} (inherits from \code{savvy_Ropendal__sealed}) of length 2.
}
\usage{
OpendalFs
}
\description{
Filesystem handle backed by Apache OpenDAL.
}
\keyword{datasets}
