% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000-wrappers.R
\docType{data}
\name{OpendalHttpFixture}
\alias{OpendalHttpFixture}
\title{Internal HTTP fixture for tests.}
\format{
An object of class \code{Ropendal::OpendalHttpFixture__bundle} (inherits from \code{savvy_Ropendal__sealed}) of length 1.
}
\usage{
OpendalHttpFixture
}
\description{
Internal HTTP fixture for tests.
}
\keyword{datasets}
