% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000-wrappers.R
\docType{data}
\name{OpendalLsIter}
\alias{OpendalLsIter}
\title{Streaming listing iterator over one prefix.}
\format{
An object of class \code{Ropendal::OpendalLsIter__bundle} (inherits from \code{savvy_Ropendal__sealed}) of length 0.
}
\usage{
OpendalLsIter
}
\description{
Streaming listing iterator over one prefix.
}
\keyword{datasets}
