% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000-wrappers.R
\docType{data}
\name{OpendalReadIter}
\alias{OpendalReadIter}
\title{Chunked read iterator over one object.}
\format{
An object of class \code{Ropendal::OpendalReadIter__bundle} (inherits from \code{savvy_Ropendal__sealed}) of length 0.
}
\usage{
OpendalReadIter
}
\description{
Chunked read iterator over one object.
}
\keyword{datasets}
