% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000-wrappers.R
\docType{data}
\name{OpendalWriteIter}
\alias{OpendalWriteIter}
\title{Chunked write sink for one object.}
\format{
An object of class \code{Ropendal::OpendalWriteIter__bundle} (inherits from \code{savvy_Ropendal__sealed}) of length 0.
}
\usage{
OpendalWriteIter
}
\description{
Chunked write sink for one object.
}
\keyword{datasets}
